old-tests/submission/W3C/domhtml/HTMLOptionElement07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OPTION</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-option-element.html#dom-option-label" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post" name="form1">
<p>
<select dir="ltr" id="selectId" multiple="multiple" name="select1" size="1" tabindex="7">
<option selected="selected" value="10001">EMP10001</option>
<option label="l1">EMP10002</option>
<option>EMP10003</option>
<option>EMP10004</option>
<option>EMP10005</option>
</select>
</p>
</form>
<p>
<select disabled="disabled" name="select2">
<option>EMP20001</option>
<option>EMP20002</option>
<option>EMP20003</option>
<option>EMP20004</option>
<option disabled="disabled">EMP20005</option>
</select>
</p>









<script>
var domtest = async_test("The label attribute is used in hierarchical menus.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vlabel;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("option");
      assert_equals(nodeList.length,10);
testNode = nodeList.item(1);
      vlabel = testNode.label;

      assert_equals(vlabel,"l1");
       
    domtest.done();
}));

</script>
<p>The label attribute is used in hierarchical menus.  It specifies          
    a shorter label for an option that the content of the OPTION element.

    Retrieve the label attribute from the second OPTION element
    and examine its value.</p>
<div id="log"></div>
</body></html>
